- SceneViewWindow.cpp: RenderSceneView 함수에 주석 추가
- ConvexMeshResource.cpp: PxDefaultMemoryOutputStream writeBuffer 추가
- DynamicRigidBody.cpp: Initialize 함수에 여러 플래그 설정 추가
- HeightFieldResource: 생성자에서 int* height를 float* height로 변경
- PhysicsCommon.h: m_convexTypeId 추가 및 구조체 수정
- PhysicsEventCallback.cpp: onContact 및 onTrigger 함수에 출력 추가
- ModelLoader.cpp: LoadModel 함수에 주석 추가 및 새로운 컴포넌트 추가
- ArticulationData.h, ArticulationLoader.h: 새로운 클래스 추가
- BoxColliderComponent, CapsuleColliderComponent: 초기화 방식 변경
- PhysicsManager.cpp: 여러 함수에서 변경 사항 적용
- RagdollComponent: 새로운 클래스 및 속성 추가
- RegisterReflect.def: 새로운 클래스 등록
- Terrain 관련 파일: TerrainCollider를 TerrainComponent로 변경
- version.txt: 커밋 해시 변경

+++ Physics/HeightFieldResource.cpp
@@ -1,7 +1,7 @@
 #include "HeightFieldResource.h"
 #include <vector>
 
-HeightFieldResource::HeightFieldResource(physx::PxPhysics* physics, const int* height, const unsigned int& numCols, const unsigned int& numRows) : ResourceBase(EResourceType::HEIGHT_FIELD)
+HeightFieldResource::HeightFieldResource(physx::PxPhysics* physics, const float* height, const unsigned int& numCols, const unsigned int& numRows) : ResourceBase(EResourceType::HEIGHT_FIELD)
 {
 	//height field 데이터 생성
 	std::vector<physx::PxHeightFieldSample> samples(numRows * numCols);
